+++ tests/testthat/fixtures/script.R
@@ -1,0 +1,3 @@
+
+cat("stdout\n")
+message("stderr")